* update labler to match the repo labels

* Avoid choking Postgres with too-large input (#1925)

Since passing the `keys` list to Postgres seems to be exceeding some
limit on the size of an SQL query, let's just ask it for all project IDs
and do the filtering afterwards if the `keys` list is large (> 100).

* Add confirmation for project deletion (#1928)

* Add dangerous actions handling to delete dialog

- Update `promptDelete` in DialogsService to include a `dangerous` flag.
- Modify delete dialog to require explicit confirmation for dangerous actions.
- Improve Sandbox component with examples for delete dialog usage.

* Confirm before deleting a project

* Restructure delete-dialog options and add link to sync-status

---------


* Remove UTF-8 BOM from all files except .verified.* (#1926)

The .verified.* files are compared to output that will include a BOM, so
they need to be left alone.

* gitignore project-cache.json (#1936)

* update git blame ignore revs

---------

+++ frontend/viewer/src/home/HomeView.svelte
@@ -28,6 +28,8 @@
   import {cubicOut} from 'svelte/easing';
   import {transitionContext} from './transitions';
   import Anchor from '$lib/components/ui/anchor/anchor.svelte';
+  import DeleteDialog from '$lib/entry-editor/DeleteDialog.svelte';
+  import {SYNC_DIALOG_QUERY_PARAM} from '../project/SyncDialog.svelte';
 
   const projectsService = useProjectsService();
   const importFwdataService = useImportFwdataService();
@@ -45,9 +47,9 @@
       .replace(/-$/, '');
   }
 
-  let customExampleProjectName: string = '';
+  let customExampleProjectName = $state('');
 
-  let createProjectLoading = false;
+  let createProjectLoading = $state(false);
 
   async function createExampleProject() {
     try {
@@ -67,10 +69,16 @@
     }
   }
 
-  let deletingProject: undefined | string = undefined;
+  let deletingProject = $state<string>();
 
   async function deleteProject(project: IProjectModel) {
+    if (!deleteDialog) throw new Error('Delete dialog not initialized');
     try {
+      const syncDialogUrl = `/project/${project.code}/browse?${SYNC_DIALOG_QUERY_PARAM}=true`;
+      if (!await deleteDialog.prompt($t`Project`, $t`${project.name}`, {
+        isDangerous: true,
+        details: $t`Make sure your [changes are synced](${syncDialogUrl}) to Lexbox.`,
+      })) return;
       deletingProject = project.id;
       await projectsService.deleteProject(project.code);
       await refreshProjects();
@@ -79,7 +87,7 @@
     }
   }
 
-  let importing = '';
+  let importing = $state('');
 
   async function importFwDataProject(name: string) {
     if (importing) return;
@@ -92,9 +100,9 @@
     }
   }
 
-  let projectsPromise = projectsService
+  let projectsPromise = $state(projectsService
     .localProjects()
-    .then((projects) => projects.sort((p1, p2) => p1.name.localeCompare(p2.name)));
+    .then((projects) => projects.sort((p1, p2) => p1.name.localeCompare(p2.name))));
 
   async function refreshProjects() {
     let promise = projectsService.localProjects().then((p) => p.sort((p1, p2) => p1.name.localeCompare(p2.name)));
@@ -103,7 +111,7 @@
   }
 
   const supportsTroubleshooting = useTroubleshootingService();
-  let troubleshootDialog: TroubleshootDialog | undefined;
+  let troubleshootDialog = $state<TroubleshootDialog>();
 
   let clickCount = 0;
   let clickTimeout: ReturnType<typeof setTimeout> | undefined;
@@ -125,8 +133,10 @@
       clickTimeout = setTimeout(resetClickCounter, 500);
     }
   }
-
+  let deleteDialog = $state<DeleteDialog>();
 </script>
+
+<DeleteDialog bind:this={deleteDialog}/>
 
 <AppBar tabTitle={$t`Dictionaries`}>
   {#snippet title()}
@@ -216,11 +226,9 @@
                         {$t`Troubleshoot`}
                       </ResponsiveMenu.Item>
                     {/if}
-                    {#if $isDev}
-                      <ResponsiveMenu.Item icon="i-mdi-delete" onSelect={() => void deleteProject(project)}>
-                        {$t`Delete`}
-                      </ResponsiveMenu.Item>
-                    {/if}
+                    <ResponsiveMenu.Item icon="i-mdi-delete" onSelect={() => void deleteProject(project)}>
+                      {$t`Delete`}
+                    </ResponsiveMenu.Item>
                   </ResponsiveMenu.Content>
                 </ResponsiveMenu.Root>
               </div>
